+++ db/migrate/20170607165804_add_comments_count_to_sheets.rb
@@ -1,0 +1,6 @@
+class AddCommentsCountToSheets < ActiveRecord::Migration[5.1]
+  def change
+    add_column :sheets, :comments_count, :integer, null: false, default: 0
+    add_index :sheets, :comments_count
+  end
+end